Optimized some functions and made them descriptive
Hello, thanks for this repo! I have been using this repo since a couple of months (mainly on segmented nuclei from H&E slides), and thought of contributing to it to make it a bit better. 
If you are interested, I can also add a StarDist2D backend, in addition to CellPose.